requirements.txt
altair==4.2.0
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.5
bqplot==0.12.36
branca==0.6.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colour==0.1.5
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
filelock==3.8.2
folium==0.13.0
fonttools==4.38.0
gdown==4.6.0
geojson==2.5.0
gitdb==4.0.10
GitPython==3.1.29
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.17.1
ipyleaflet==0.17.2
ipython==8.7.0
ipytree==0.2.2
ipywidgets==8.0.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
leafmap==0.14.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pyshp==2.3.1
PySocks==1.7.1
pystac==1.6.1
pystac-client==0.5.1
python-box==6.1.0
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
scooby==0.7.0
seaborn==0.12.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
stack-data==0.6.2
streamlit==1.15.2
streamlit-folium==0.7.0
streamlit-metrics==0.1.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
traittypes==0.2.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.0
wcwidth==0.2.5
whitebox==2.2.0
whiteboxgui==2.2.0
widgetsnbextension==4.0.3
xyzservices==2022.9.0
zipp==3.11.0
